Update baseclient.py
Reliability has seen some sockets not get shutdown correctly when investigating magic number faults. The client code gives no indication that we faulted on shutdown/close, but we should probably alert the user if we weren't able to close the socket in a clean fashion. This change should address that.